test(abtesting/hooks): move tests to the root
