+++ components/abtesting/hooks/test/optimizely-x.test.js
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-import Service from '../useExperimentCore/optimizely-x'
+import Service from '../src/useExperimentCore/optimizely-x'
 
 const fakeExperiments = {
   100: {
